+++ lib/src/test/java/com/quiltdata/quiltcore/WorkflowsTest.java
@@ -23,11 +23,11 @@ import com.quiltdata.quiltcore.workflows.WorkflowValidator;
 public class WorkflowsTest {
     @Test
     public void testConfigDataVersion() {
-        assertEquals(ConfigDataVersion.parse("1.2.3"), new ConfigDataVersion(1, 2, 3));
-        assertEquals(ConfigDataVersion.parse("1.2"),   new ConfigDataVersion(1, 2, 0));
-        assertEquals(ConfigDataVersion.parse("1"),     new ConfigDataVersion(1, 0, 0));
+        assertEquals(new ConfigDataVersion(1, 2, 3), ConfigDataVersion.parse("1.2.3"));
+        assertEquals(new ConfigDataVersion(1, 2, 0), ConfigDataVersion.parse("1.2"));
+        assertEquals(new ConfigDataVersion(1, 0, 0), ConfigDataVersion.parse("1"));
 
-        assertTrue(ConfigDataVersion.parse("1.2.3").compareTo(ConfigDataVersion.parse("1.2.3")) == 0);
+        assertEquals(0, ConfigDataVersion.parse("1.2.3").compareTo(ConfigDataVersion.parse("1.2.3")));
         assertTrue(ConfigDataVersion.parse("1.2.3").compareTo(ConfigDataVersion.parse("1.2.2")) > 0);
         assertTrue(ConfigDataVersion.parse("1.2.3").compareTo(ConfigDataVersion.parse("1.3")) < 0);
         assertTrue(ConfigDataVersion.parse("1.2").compareTo(ConfigDataVersion.parse("2")) < 0);
@@ -40,8 +40,7 @@ public class WorkflowsTest {
         PhysicalKey key = new LocalPhysicalKey(path);
         WorkflowConfig config = WorkflowConfig.load(key);
 
-        assertNotEquals(null, config);
-
+        assert config != null;
         WorkflowValidator validator = config.getWorkflowValidator("alpha");
 
         JsonNode data = validator.getDataToStore();
@@ -60,23 +59,15 @@ public class WorkflowsTest {
         validator.validate("test/foo", Map.of("README.md", readme), pkgMeta, "blah");
 
         // Bad name.
-        assertThrows(WorkflowException.class, () -> {
-            validator.validate("zzz/foo", Map.of("README.md", readme), pkgMeta, "blah");
-        });
+        assertThrows(WorkflowException.class, () -> validator.validate("zzz/foo", Map.of("README.md", readme), pkgMeta, "blah"));
 
         // Missing commit message.
-        assertThrows(WorkflowException.class, () -> {
-            validator.validate("test/foo", Map.of("README.md", readme), pkgMeta, null);
-        });
+        assertThrows(WorkflowException.class, () -> validator.validate("test/foo", Map.of("README.md", readme), pkgMeta, null));
 
         // Missing README.
-        assertThrows(WorkflowException.class, () -> {
-            validator.validate("test/foo", Map.of(), pkgMeta, "blah");
-        });
+        assertThrows(WorkflowException.class, () -> validator.validate("test/foo", Map.of(), pkgMeta, "blah"));
 
         // Wrong package metadata.
-        assertThrows(WorkflowException.class, () -> {
-            validator.validate("test/foo", Map.of("README.md", readme), readmeMeta, "blah");
-        });
+        assertThrows(WorkflowException.class, () -> validator.validate("test/foo", Map.of("README.md", readme), readmeMeta, "blah"));
     }
 }